Add WishlistController and WishlistRequest for managing user wishlists
