+++ app/Http/Controllers/Api/V1/ProductController.php
@@ -5,7 +5,6 @@ namespace App\Http\Controllers\Api\V1;
 use App\Http\Controllers\Controller;
 use App\Models\Product;
 use App\Http\Resources\ProductResource;
-use App\Http\Requests\ProductRequest;
 
 class ProductController extends Controller
 {
